familytree/requirements.txt
requests
djangorestframework
django_login_history
django-cors-headers